+++ CHANGELOG.markdown
@@ -3,6 +3,10 @@
 
   * Added `prefixed`.
 
+  * Added `parseOr`, a slightly generalized version of `parse`.
+
+  * Renamed `parseTest` to `parsePure`
+
 0.1.0
 =====